Add triples to sample data
This commit adds the following triples to each sample document:

<docUri> <rdfs:label> <doc/name> .
<docUri> <rdf:type> <foaf:Person> .

and three triples pointing to 3 random documents:

<docUri> <foaf:knows> <someRandomDocUriFromThisSampleSet>

There is also a single additional triple in the first doc
('data-0.json') that establishes a label for `foaf:knows`. This would
normally go into an ontology, but it seemed easiest just to embed it
into a document rather than introduce that much complexity. This
additional triple is needed for a nice OOTB display when using
[ml-visjs-graph-ng](https://github.com/patrickmcelwee/ml-visjs-graph-ng)

These triples were inserted into the original sample docs using [this
sample-triple-generator
library](https://github.com/patrickmcelwee/sample-triple-generator), in
case that is a helpful starting point for any future additions.