Hide 'Edit' and 'Delete' when `disallowUpdates: true`
Previously, the buttons still appeared, but resulted in a blank page for
'Edit' and an unhandled server error for 'Delete'.

This requires added the userService and $scope as additional
dependencies in the generated detail controller.

It seems that only the map theme has an override for the detail
controller. I made the changes there as well.