Do not add padding to .row in the dashboard theme
In Bootstrap, .row has no padding (and in fact has a negative margin)
to allow for nested columns. I just ran into this setting on a
dashboard-style demo, where I could not figure out why my nested columns
keep getting more and more indented.

I believe that this line of css was added to fix the search screen for
the dashboard theme. But a better fix is to use the "content" class
provided by AdminLTE. This is what I have done on my own demo, and it is
working well. It is also what the AdminLTE site does (though it is oddly
undocumented): https://almsaeedstudio.com/themes/AdminLTE/index2.html